# templates/factory/services/backend/openapi/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BackendService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/login": {
      "post": {
        "operationId": "BackendService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customersLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BackendService"
        ]
      }
    },
    "/organization": {
      "get": {
        "summary": "Organization for the authenticated user",
        "operationId": "BackendService_GetOrganization",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customersOrganization"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BackendService"
        ]
      },
      "post": {
        "summary": "Organization for the authenticated user",
        "operationId": "BackendService_UpdateOrganization",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customersUpdateOrganizationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/customersUpdateOrganizationRequest"
            }
          }
        ],
        "tags": [
          "BackendService"
        ]
      }
    },
    "/register": {
      "post": {
        "operationId": "BackendService_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customersRegisterUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BackendService"
        ]
      }
    },
    "/version": {
      "get": {
        "operationId": "BackendService_Version",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customersVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BackendService"
        ]
      }
    }
  },
  "definitions": {
    "customersLoginResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/customersUser"
        }
      }
    },
    "customersOrganization": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "customersRegisterUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/customersUser"
        },
        "status": {
          "$ref": "#/definitions/customersUserStatus"
        },
        "organization": {
          "$ref": "#/definitions/customersOrganization"
        }
      }
    },
    "customersUpdateOrganizationRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "domain": {
          "type": "string"
        }
      }
    },
    "customersUpdateOrganizationResponse": {
      "type": "object",
      "properties": {
        "org": {
          "$ref": "#/definitions/customersOrganization"
        }
      }
    },
    "customersUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "signupAuthId": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "profile": {
          "$ref": "#/definitions/customersUserProfile"
        },
        "signedUpAt": {
          "type": "string",
          "format": "date-time"
        },
        "lastLoginAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "customersUserProfile": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "givenName": {
          "type": "string"
        }
      }
    },
    "customersUserStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        }
      }
    },
    "customersVersionResponse": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}